hubble: move AssertProtoEqual helper to hubble testutils
This commit moves the AssertProtoEqual test helper out of the
test/helpers package which otherwise contains the test helpers for
Ginkgo-based tests. The existing pkg/hubble/testutils package seems like
a more appropriate place for the helper. Also this reduces initial test
build time for pkg/hubble from ~28s to ~15s on my machine.
