+++ pkg/hubble/testutils/proto.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package helpers
+package testutils
 
 import (
 	"fmt"
